+++ spec/fixtures/unclosed.js
@@ -1,0 +1,5 @@
+//#if FOO
+'foo'
+//#else
+'bar'
+//endif